+++ CHANGELOG.md
@@ -35,6 +35,9 @@
   - Fixed formatting issues for the function to work with modules other than SqlServer.
 - New DSC resource SqlServerSecureConnection
   - New resource to configure a SQL Server instance for encrypted SQL connections.
+- Changes to SqlAlwaysOnService
+  - Updated integration tests to use NetworkingDsc
+    ([issue #1129](https://github.com/PowerShell/SqlServerDsc/issues/1129)).
 
 ## 12.0.0.0